pyproject.toml
[project]
name = "consistency-checker"
version = "0.0.1"
description = "Detects internal contradictions across a corpus of documents using a two-stage NLI + LLM pipeline."
readme = "README.md"
requires-python = ">=3.11"
license = { file = "LICENSE" }
authors = [{ name = "contradictionchecker contributors" }]
dependencies = [
    "pydantic>=2.7",
    "pyyaml>=6.0",
    "pysbd>=0.3.4",
    "anthropic>=0.39",
    "numpy>=1.26",
    "faiss-cpu>=1.7",
    "sentence-transformers>=3.0",
]

[project.scripts]
consistency-check = "consistency_checker.cli.main:app"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["consistency_checker"]

[dependency-groups]
dev = [
    "pytest>=8.0",
    "pytest-cov>=5.0",
    "ruff>=0.6",
    "mypy>=1.10",
    "types-PyYAML>=6.0",
]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "B", "UP", "SIM", "RUF"]
ignore = ["E501"]

[tool.mypy]
python_version = "3.11"
strict = true
files = ["consistency_checker"]

[[tool.mypy.overrides]]
module = [
    "pysbd",
    "pysbd.*",
    "faiss",
    "faiss.*",
    "sentence_transformers",
    "sentence_transformers.*",
    "transformers",
    "transformers.*",
]
ignore_missing_imports = true

[tool.pytest.ini_options]
testpaths = ["tests"]
markers = [
    "slow: tests that download HF models or are otherwise expensive",
    "live: tests that hit external APIs (require API keys)",
    "e2e_fixture: end-to-end tests using fixture LLM (no network)",
]
addopts = "-ra -q --strict-markers"
filterwarnings = [
    "ignore::DeprecationWarning:pysbd.*",
    "ignore:builtin type Swig.*:DeprecationWarning",
]
